fix: rename data constants file
